Рефакторинг - перенес отрисовку в модуль отрисовки
