+++ js/main.js
@@ -1,13 +1,13 @@
 
-//import { getMockPhotosArray } from './mock-data.js';
-import { setValidationEventListeners } from './form-validation-helper.js';
-import { setImgScaleEventListeners } from './img-scale-evt-helper.js';
+import { setValidationEvt } from './form-validation-helper.js';
+import { setImgScaleEvt } from './img-scale-evt-helper.js';
 import { setEffectFieldsetEvt } from './nouislider-evt-helper.js';
-import {sendData, renderPhotos, receivedPhotos } from './fetch-helper.js';
+import {sendData, receivedPhotos } from './fetch-helper.js';
 import {setFilterButtonsEvt} from './filters-evt-helper.js';
+import {renderPhotos} from './draw-thumbnails.js';
 
 renderPhotos(receivedPhotos);
-setValidationEventListeners(sendData);
-setImgScaleEventListeners();
+setValidationEvt(sendData);
+setImgScaleEvt();
 setEffectFieldsetEvt();
 setFilterButtonsEvt();
